Add mechanism to rate the application
Add a new method to the object representing the game. It will send the user to a screen where the application can be rated.
Add another method that will determine if the user should be asked to rate the app. This depends on whether the user has already done that and the number of games played so far.
From now on, we will store a flag as game data, representing whether user already clicked the link to "rate the app".